refactor(ios): containsKorean ×5, nilIfBlank ×2 중복 → StringExtensions.swift 통합
5개 ViewModel/뷰에 복붙된 동일 String.containsKorean 확장과 2곳의 Optional<String>.
nilIfBlank 확장을 단일 출처로 통합. 호출부(.containsKorean/.nilIfBlank)는 그대로 해소.
동작 변경 없음.
