+++ apps/ios-native/VoiceAlarm/Services/IAP/SubscriptionManager.swift
@@ -277,12 +277,3 @@ enum PurchaseResult: Equatable {
     }
 }
 
-private extension String {
-    var containsKorean: Bool {
-        contains { character in
-            character.unicodeScalars.contains { scalar in
-                (0xAC00...0xD7A3).contains(Int(scalar.value))
-            }
-        }
-    }
-}
