+++ apps/ios-native/VoiceAlarm/SocialFeatureViewModel.swift
@@ -834,12 +834,3 @@ final class SocialFeatureViewModel: ObservableObject {
     }
 }
 
-private extension String {
-    var containsKorean: Bool {
-        contains { character in
-            character.unicodeScalars.contains { scalar in
-                (0xAC00...0xD7A3).contains(Int(scalar.value))
-            }
-        }
-    }
-}
