+++ apps/ios-native/VoiceAlarm/VoiceStudioViewModel+ErrorMapping.swift
@@ -109,12 +109,3 @@ extension VoiceStudioViewModel {
     ]
 }
 
-private extension String {
-    var containsKorean: Bool {
-        contains { character in
-            character.unicodeScalars.contains { scalar in
-                (0xAC00...0xD7A3).contains(Int(scalar.value))
-            }
-        }
-    }
-}
